Refactor entry signing, encoding and decoding logic into own module
